selectors/selectors/selectors-normalization-001.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Selectors: markup precomposed, CSS precomposed</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/selectors/#lang-pseudo'>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel="stylesheet" type="text/css" href="../../resources/testharness.css">
<meta name="assert" content="[Exploratory test] A fully precomposed class name will match a fully precomposed CSS selector name.">
<style type='text/css'>
.test div { width: 50px; }
#box.ti\1EBF ngvi\1EC7 t { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" class="ti&#x1EBF;ngvi&#x1EC7;t">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>

<div id='log'></div>

<!-- Notes:

This test should always work.

-->
</body>
</html>
